Upate build-restricter.sh for new AMD-ified world.
This is a good demonstration of using the requirejs optimizer (r.js) to
pack the entire 'restricter' code into a single file.  We also automatically
test that the code base runs correctly in 'restrict mode' and generates
identical output.

This commit adds a new shaper, "deeval", which replaces the 'eval' statements
in the Narcissus code base in order to make Narcissus "use strict" safe.