Be more careful with delimiters when using Shaper._insert.
When adding last child and the comma is (for example) a COMMA, we don't
usually want to add the trailing delimiter.  But when the delimiter is
whitespace, we usually want to always add the delimiter.  Add an extra
optional parameter to distinguish these two cases (and fix the "comma
after last element in list" case).